requirements.txt
requests~=2.32.3
fastapi~=0.115.6
pydantic~=2.10.4
python-jose~=3.3.0
python-dotenv~=1.0.1
aiogram~=3.17.0